[BREAKING CHANGE] Adding support for CRUD Folders, Users (Partial) DataSource Filters,
ChangeLog:
  - Folders CRUD operations
  - Import, Clear added for users.
  - TechDebt: Flattening cmd list to single file for all operations.
  - DataSource Updates related to configuration, this is a breaking
    change.  See updated config examples and documentation.
  - Adding documentation for installation and configuration changes

Fixes #88, #91, #84
