--- cmd/contextClear.go
@@ -1,21 +1,0 @@
-package cmd
-
-import (
-	"github.com/esnet/gdg/apphelpers"
-	log "github.com/sirupsen/logrus"
-	"github.com/spf13/cobra"
-)
-
-var contextClear = &cobra.Command{
-	Use:   "clear",
-	Short: "clear all context",
-	Long:  `clear all contexts`,
-	Run: func(cmd *cobra.Command, args []string) {
-		apphelpers.ClearContexts()
-		log.Info("Successfully deleted all configured contexts")
-	},
-}
-
-func init() {
-	context.AddCommand(contextClear)
-}
