--- cmd/devel_completion.go
@@ -1,38 +1,0 @@
-package cmd
-
-import (
-	log "github.com/sirupsen/logrus"
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-var CompletionCmd = &cobra.Command{
-	Use:                   "completion [bash|zsh|fish|powershell]",
-	Short:                 "Generate completion script",
-	Long:                  "To load completions",
-	DisableFlagsInUseLine: true,
-	ValidArgs:             []string{"bash", "zsh", "fish", "powershell"},
-	Args:                  cobra.ExactValidArgs(1),
-	Run: func(cmd *cobra.Command, args []string) {
-		var err error
-		switch args[0] {
-		case "bash":
-			err = cmd.Root().GenBashCompletion(os.Stdout)
-		case "zsh":
-			err = cmd.Root().GenZshCompletion(os.Stdout)
-		case "fish":
-			err = cmd.Root().GenFishCompletion(os.Stdout, true)
-		case "powershell":
-			err = cmd.Root().GenPowerShellCompletion(os.Stdout)
-		}
-		if err != nil {
-			log.Error("Failed to generation shell completion.")
-		}
-	},
-}
-
-func init() {
-	develCmd.AddCommand(CompletionCmd)
-
-}
