+++ cmd/root.go
@@ -45,6 +45,7 @@ func Execute() {
 
 func init() {
 	cobra.OnInitialize(initConfig)
+	rootCmd.CompletionOptions.DisableDefaultCmd = true
 	rootCmd.PersistentFlags().StringP("config", "c", "", "Configuration Override")
 }
