+++ documentation/content/docs/usage_guide.md
@@ -5,6 +5,17 @@ weight: 14
 
 Every namespace supporting CRUD operations has the functions: list, import, export, clear operating on only the monitored folders.
 
+### Alert Notifications
+
+Allows you to manage alertnotifications (an) if you have any setup
+
+```sh
+./bin/gdg an list -- Lists all alert notifications
+./bin/gdg an import -- retrieve and save all alertnotifications from grafana
+./bin/gdg an export  -- writes all local alert notifications to grafana
+./bin/gdg an clear -- Deletes all alert notifications
+```
+
 ### Contexts
 
 Starting with version 0.1.4 contexts are now supported.  Your config can contain one or multiple contexts which are essentially a grafana server configuration.
@@ -19,6 +30,7 @@ ctx is shorthand for context and basic CRUD is supported which is mainly tooling
 ./bin/gdg ctx cp qa staging -- copies the qa context to staging and sets it as active
 ./bin/gdg ctx clear -- Will delete all active contexts leaving only a single example entry
 ```
+
 
 #### Dashboards
 
@@ -48,6 +60,26 @@ All commands can use `datasources` or `ds` to manage datasources
 ./bin/gdg ds clear -- Deletes all datasources
 ```
 
+#### Devel
+Some developer helper utilities
+
+
+```sh 
+./bin/gdg devel completion  [bash|fish|powershell|zsh] --  Will generate autocompletion for GDG for your favorite shell
+./bin/gdg devel srvinfo -- print grafana server info
+```
+
+#### Folders
+
+Mostly optional as Dashboards will create/delete these are needed but if there is additional metadata you wish to persist you can use this to manage them.
+
+```sh
+./bin/gdg folders list -- Lists all current folders
+./bin/gdg folders import -- Import all folders from grafana to local file system
+./bin/gdg folders export -- Exports all folders from local filesystem 
+./bin/gdg folders clear -- Deletes all folders
+```
+
 #### Organizations
 Command can use `organizations` or `org` to manage organizations.
 
@@ -59,8 +91,29 @@ Command can use `organizations` or `org` to manage organizations.
 
 Only supported with basic auth.  Users is the only one where basic auth is given priority.  API Auth is not supported, so will try to use basic auth if configured otherwise will warn the user and exit.
 
+NOTE: admin user is always ignored.  
+
 ```sh
 ./bin/gdg users list -- Lists all known users
 ./bin/gdg users promote -u user@foobar.com -- promotes the user to a grafana admin
+./bin/gdg users import -- Lists all known users
+./bin/gdg users export -- Export all users (Not yet supported)
+./bin/gdg users clear -- Delete all known users except admin
+```
+### Version
+
+Print the applications release version
+
+```sh 
+./bin/gdg version
 ```
 
+
+```
+Build Date: 2022-05-05-13:27:08
+Git Commit: 34cc84b3d80080aa93e74ed37739bddc3638997c+CHANGES
+Version: 0.1.11
+Go Version: go1.18
+OS / Arch: darwin amd64
+
+```
